drivers: add support for out-of-tree drivers
An external project extending the Zephyr RTOS and its drivers may have
subsystems that needs to use its own specific driver. An example is the
nRF5x NVMC that needs to be scheduled in between radio operations.

In order to not introduce non-Zephyr specific code into the Zephyr tree,
an option is added to disable the default drivers in Zephyr. The
out-of-tree subsystem will select, e.g., FLASH_HAS_SUBSYS_DRIVER.

Support for out-of-tree drivers is added to clock_control, entropy and
flash.

Signed-off-by: Thomas Stenersen <thomas.stenersen@nordicsemi.no>